Backend/requirements-prod.txt
# ===== PRODUCTION REQUIREMENTS =====
# Incluir requirements principales
-r requirements.txt

# ===== PRODUCTION SERVER =====
gunicorn==21.2.0
uvicorn==0.24.0
gevent==23.9.1

# ===== DATABASE PRODUCTION =====
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# ===== CACHING & SESSIONS =====
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3

# ===== STATIC FILES & MEDIA =====
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
botocore==1.34.0

# ===== MONITORING & LOGGING =====
sentry-sdk==1.38.0
newrelic==9.3.0
structlog==23.2.0
python-json-logger==2.0.7

# ===== SECURITY PRODUCTION =====
django-security==0.17.0
django-ratelimit==4.1.0
django-csp==3.7
django-permissions-policy==4.18.0

# ===== PERFORMANCE =====
django-cachalot==2.6.1
django-compression-middleware==0.4.1

# ===== SSL & HTTPS =====
certifi==2023.11.17
urllib3==2.1.0

# ===== ENVIRONMENT =====
python-dotenv==1.0.0
environs==10.3.0